Move files to a different package
